requirements.txt
absl-py==1.4.0
ale-py==0.7.4
alembic==1.10.2
astunparse==1.6.3
atomicwrites==1.4.1
attrs==18.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.5.5
box2d-py==2.3.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.25.2
colorama==0.4.3
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
DataProperty==0.55.0
decorator==4.4.2
derivative==0.5.3
dm-tree==0.1.8
fasteners==0.18
filelock==3.9.1
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
glfw==2.5.6
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==3.0.0a6
gradient==2.0.6
gradient-utils==0.5.0
graphql-core==3.1.7
greenlet==2.0.2
grpcio==1.51.3
gym==0.21.0
gym-notices==0.0.8
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h5py==3.8.0
halo==0.0.31
huggingface-hub==0.13.2
huggingface-sb3==2.2.4
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
INNLab==0.9.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
libtorrent==2.0.7
log-symbols==0.0.14
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==2.21.0
matplotlib==3.7.0
mbstrdecoder==1.1.2
mdurl==0.1.2
moviepy==1.0.3
mujoco==2.2.0
mujoco-py==2.1.2.14
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optuna==3.1.0
packaging==23.0
pandas==1.5.3
pathvalidate==2.5.2
Pillow==9.4.0
pluggy==1.0.0
proglog==0.1.10
progressbar2==4.2.0
prometheus-client==0.9.0
protobuf==3.19.6
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
Pygments==2.14.0
pymongo==3.13.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pysindy==1.7.3
pytablewriter==0.64.2
pytest==7.0.1
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.2
rl-zoo3==1.7.0
rsa==4.9
sb3-contrib==1.7.0
scikit-learn==1.0.2
scipy==1.10.1
shellingham==1.5.0.post1
six==1.16.0
spinners==0.0.24
SQLAlchemy==2.0.6
stable-baselines3==1.7.0
swig==4.1.1
tabledata==1.3.1
tcolorpy==0.1.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensordict==0.0.2b0
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchrl==0.0.4
torchvision==0.14.1
tqdm==4.64.1
typepy==1.3.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
websocket-client==0.57.0
Werkzeug==2.2.3
wincertstore==0.2
windows-curses==2.3.1
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
